Further API updates

+++ OpenIIoT.SDK.Packaging/Operations/PackageVerifier.cs
@@ -126,11 +126,6 @@ namespace OpenIIoT.SDK.Packaging.Operations
         {
             ArgumentValidator.ValidatePackageFileArgumentForReading(packageFile);
 
-            if (!string.IsNullOrEmpty(publicKey))
-            {
-                ArgumentValidator.ValidatePublicKeyArgument(publicKey);
-            }
-
             Info($"Verifying Package '{Path.GetFileName(packageFile)}'...");
 
             Exception deferredException = default(Exception);
@@ -217,9 +212,13 @@ namespace OpenIIoT.SDK.Packaging.Operations
             }
             finally
             {
-                Verbose("Deleting temporary files...");
-                Directory.Delete(tempDirectory, true);
-                Verbose("Temporary files deleted successfully.");
+                if (Directory.Exists(tempDirectory))
+                {
+                    Verbose("Deleting temporary files...");
+
+                    Directory.Delete(tempDirectory, true);
+                    Verbose("Temporary files deleted successfully.");
+                }
 
                 if (deferredException != default(Exception))
                 {
